Use new DER decoding and encoding routines.
 * Benefits are less copies of sensitive data in memory.
 * More insightful parsing and API.

+++ pkcs11/gkm/gkm-certificate-trust.c
@@ -83,7 +83,7 @@ has_key_usage (GkmCertificateTrust *self, guint check, CK_ULONG *val)
 	GkmDataResult res;
 	const guchar *extension;
 	gsize n_extension;
-	guint usage;
+	gulong usage;
 
 	g_return_val_if_fail (self->pv->certificate, CKR_GENERAL_ERROR);
 	*val = CKT_NETSCAPE_TRUST_UNKNOWN;
